Feature | Teacher can see his cohorts

+++ app/Http/Controllers/CohortController.php
@@ -9,26 +9,38 @@ use Illuminate\Contracts\View\Factory;
 use Illuminate\Contracts\View\View;
 use Illuminate\Foundation\Application;
 use Illuminate\Http\Request;
-
-use Carbon\Carbon;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 
 class CohortController extends Controller
 {
+    use AuthorizesRequests;
     /**
      * Display all available cohorts
      * @return Factory|View|Application|object
      */
     public function index() {
 
-        // Récupérer les promotions avec écoles + étudiants
-        $cohorts = Cohort::with('school', 'students')->get();
+        $user = auth()->user(); // Récupérer l'utilisateur actuellement connecté
 
+        // Vérifier si l'utilisateur a le droit de voir les cohortes avec la méthode viewAny de la policy
+        $this->authorize('viewAny', Cohort::class);
 
+        // Si l'utilisateur est un admin, on récupère toutes les cohortes
+        if ($user->school()->pivot->role == 'admin') {
+            // Récupérer toutes les cohortes avec écoles et étudiants
+            $cohorts = Cohort::with('school', 'students')->get();
+        } else {
+            // Si l'utilisateur est un teacher, on récupère uniquement les cohortes auxquelles il est associé
+            $cohorts = $user->cohorts() // On utilise la relation définie dans le modèle User pour les cohortes
+            ->with('school', 'students')
+                ->get();
+        }
 
         // Récupérer les écoles pour le dropdown
         $schools = School::pluck('name', 'id');
 
+        // Retourner la vue avec les cohortes et les écoles
         return view('pages.cohorts.index', compact('cohorts', 'schools'));
     }
 
@@ -40,27 +52,39 @@ class CohortController extends Controller
      */
     public function show(Cohort $cohort) {
 
+
+
         $students = User::join('users_schools', 'users.id', '=', 'users_schools.user_id')
             ->where('users_schools.role', 'student')
             ->select('users.*')
             ->get();
 
-
+        // Récupérer tous les enseignants (role = 'teacher')
+        $teachers = User::join('users_schools', 'users.id', '=', 'users_schools.user_id')
+            ->where('users_schools.role', 'teacher')
+            ->select('users.*')
+            ->get();
 
         $cohortStudents = $cohort->students;
+        $cohortTeachers = $cohort->teachers;
 
 
         return view('pages.cohorts.show', [
             'cohort' => $cohort,
             'students' => $students,
-            'cohortStudents' => $cohortStudents
-        ]);
+            'cohortStudents' => $cohortStudents,
+            'teachers' => $teachers,
+            'cohortTeachers' => $cohortTeachers,
+            ]);
+
     }
 
 
 
     public function store(Request $request)
     {
+        $this->authorize('create', Cohort::class);
+
         // Validation
         $validated = $request->validate([
             'name' => 'required|string|max:255',
@@ -81,6 +105,8 @@ class CohortController extends Controller
         // Récupérer l'ID
         $cohort = Cohort::findOrFail($id);
 
+        $this->authorize('delete', $cohort);
+
         // Supprimer la cohorte
         $cohort->delete();
 
@@ -91,6 +117,8 @@ class CohortController extends Controller
     {
         // Trouver l'ID
         $cohort = Cohort::findOrFail($id);
+
+        $this->authorize('update', $cohort);
 
         // Validation
         $validatedData = $request->validate([
@@ -117,6 +145,8 @@ class CohortController extends Controller
 
     public function addStudent(Request $request, Cohort $cohort)
     {
+        $this->authorize('addStudent', $cohort);
+
         // Valider l'ID de l'étudiant
         $validated = $request->validate([
             'user_id' => 'required|exists:users,id',
@@ -131,10 +161,39 @@ class CohortController extends Controller
     public function deleteStudent(Cohort $cohort, $userId)
     {
 
+        $this->authorize('deleteStudent', $cohort);
+
         $cohort->students()->detach($userId);
         return redirect()->route('cohort.show', $cohort);
 
     }
+
+    public function addTeacher(Request $request, Cohort $cohort)
+    {
+        $this->authorize('addTeacher', $cohort);
+
+        // Valider l'ID de l'enseignant
+        $validated = $request->validate([
+            'user_id' => 'required|exists:users,id',
+        ]);
+
+        // Ajouter l'enseignant à la cohorte (via la table pivot cohort_teacher)
+        $cohort->teachers()->attach($validated['user_id']);
+
+        return redirect()->route('cohort.show', $cohort);
+    }
+
+    public function deleteTeacher(Cohort $cohort, $userId)
+    {
+        $this->authorize('deleteTeacher', $cohort);
+
+        // Supprimer l'enseignant de la cohorte (via la table pivot cohort_teacher)
+        $cohort->teachers()->detach($userId);
+
+        return redirect()->route('cohort.show', $cohort);
+    }
+
+
 
     public function getForm(Cohort $cohort)
     {
